Inital CloudWatch sink (#90)
This is the first pass at a cloudwatch logs sink. This one shares a lot of functionality with the s3 sink but requires a bit of a more involved request future. This is because to put logs into a cloudwatch logs stream you must get a sequence token first. This means, the first put request will have to go fetch that token, then every subsequent request can use the token returned from the previous put request.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>