Add structured values as `Bytes` and forward compat proto (#226)
This switches the structred field in record to be a `HashMap<Atom, Bytes>`. There are a few caveats with this approach, first is encoding into a string is a bit easier with the `BufExt` extension trait, but the `Bytes` must first be `IntoBuf`ed then you can call those functions on it. Second, `Bytes` can NOT be used with `json!` from `serde_json` it will get silently dropped. To get around this all you need to do is convert the `Bytes` into a `String`.

Closes #211 
Closes #185
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>