feat(aws_cloudwatch_logs): Add retry ability to cloudwatch (#663)
This PR introduces a large refactor of the cloudwatch logs sink. This refactor was needed to fix issues where the sink would panic if the put_logs request failed.

The solution is to move the entire request cycle (create stream, describe stream, put logs) into its own future. The actual state machine implementation is almost identical to the previous iteration. The only major change is now a mpsc channel is used to send the new token from the put logs response.

Cloudwatch is now broken into two files, the main module file contains both services and all the configuration. The second, request.rs file contains the future that is used to fire a cloudwatch request.

The other changes are, tests now use a non shared client to be able to parallelize tests. And tracing was updated to v0.1.2 and tower was updated to v0.1.1.

This change also includes the ability to retry cloudwatch requests and includes the new config items.

Testing wise I modified the mocker to recreate failed requests and it seems to work. If any of the requests that are not put logs fail, it will return the token so it can retry. Otherwise, it will return None since the token was consumed by the put logs request and it will restart the cycle. This is totally safe.

Closes #632, #301

Signed-off-by: Lucio Franco <luciofranco14@gmail.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>