feat(blackhole sink): Accept metric events, too (#1237)
The Vector website indicates that the blackhole sink accepts both log
and metric events, but this isn't true: it only accepts logs. There's no
reason this should be the case.

In particular, for metric sources (such as `statsd` in my case) that
want to accept/transform metrics but just send everything to `/dev/null`
for the moment (e.g. when debugging your configuration), there's no easy
way to do this, resulting in an annoying error in the startup topology
check.

I also like enabling blackholes with high `print_amount` settings
anyway, so that my `journald` logs show some progress occasionally. This
will allow me to send all streams to one blackhole that prints that
info.

This patch simply takes events and distinguish metrics, adding them to
the total event count, and taking the total length of the .to_string()
method for the byte size, AKA the json representation. I believe this is
roughly the right metric, since it is basically a measure of "how many
bytes this stream will roughly push out", which will generally be some
rendered form of the metric data -- such as JSON.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>