fix(testing): remove sleeps from topology tests (#1346)
Instead of relying on sleeps to ensure that our mock source has a chance
to process messages, we make its input an unbuffered channel. This means
that the `send` future in our test will not resolve until the source
actually has possession of the input, as opposed to it sitting in the
channel buffer. This should remove all need to worry about timing for
these interactions.

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>