docs: Fix handling of locally-installed bundles (#1354)
* Don't recurse into any installed Ruby bundles when generating docs

The docs generation script `scripts/generate.rb` recurses into all
directories to find `.erb` files to transform. However, on some systems,
the `bundle install` command in the `Makefile` will install the required
gems into a `vendor` directory within the Vector source tree. This
results in `scripts/generate.rb` attempting to generate docs in the
gems, leading to errors.

This change removes all filenames that appear to be part of a bundle
from the list of files to generate.

Signed-off-by: Bruce Guenter <bruce@timber.io>

* Rework `scripts/generate.rb` to skip all of `scripts`

The source tree no longer stores any `.erb` files in the `scripts`
directory, so just skip all of it.

Signed-off-by: Bruce Guenter <bruce@timber.io>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>