fix(loki sink): Set a proper default for the batch buffer initialization (#3016)
* Use BatchSize in Buffer instead of breaking it down

* Set a default event buffer size for the loki sink

The loki sink uses VecBuffer but did not set a default for the maximum
number of events. When unset, the default number of events is
`usize::max_value()`. Further, VecBuffer creates an array based on this
maximum number of events. This causes a run time error due to the vector
size being too large for memory.

Signed-off-by: Bruce Guenter <bruce@timber.io>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>