enhancement(splunk_hec sink): Add configuration for source (#3327)
* enhancement(splunk_hec sink): Add configuration for source

This adds a configuration option to set the `source` attribute of events
sent to Splunk. Typically this is the filename or network address where
the events came from.

https://docs.splunk.com/Splexicon:Source

Eventually, I could see us setting this automatically (maybe if they set
a sentinel value like `auto`) if the events come from the `file` source
(and possibly others as well), but this'll at least allow users to
configure it. They could set it to `{{ file }}` if the events came from
the `file` source to have it be the filename.

Fixes #3251

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>