chore: Add a test implementation of bors (#4178)
This PR enables https://bors.tech for our PR merge workflow. Now we run an abbreviated lint-centric check on the PR rather than the full test suite. We've achieved this by running the lint tests inside a container running the `timberio/ci_image` Docker image. 

We also consolidated and updated the image build process in `environment.yml` to keep the `ci_image` up-to-date.

Then, instead of pressing merge, we use the `bors` commands to submit the PR to a `staging` branch where CI is run. The new CI run merges the prior e2e and tests jobs as a group. If these return green then the `staging` branch is fast-forwarded/merged into `master`.

Signed-off-by: James Turnbull <james@lovedthanlost.net>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>