fix(metrics): Add missing methods to Subscriber forwarder (#5529)
The internal_logs source introduced in
e49c4254402b5a82bf11d4b158278376a84c6905 added a Subscriber forwarding
wrapper so it could send events to both a broadcast bus and to the
formatter. This wrapper was missing the `downcast_raw` method, which
apparently broke internal metrics. This patch adds both that method and
the `max_level_hint` which was also missing.

Signed-off-by: Bruce Guenter <bruce@timber.io>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>