fix(tests): podman pod stop/rm don't need a --name option (#5669)
Most of the `stop-integration-*` targets in the Makefile used the
`--name` option to `podman stop` and `podman rm` to specify the pod
name. This resulted in an error, since the option is not needed, but the
error was hidden by an error redirect to `/dev/null`. This patch removes
the use of that option.

Signed-off-by: Bruce Guenter <bruce@timber.io>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>