fix(influxdb_metrics sink): Fix transmission of counter metrics (#5922)
This reverts commit a8de8b42657bf77d7b58c82dd997b5aa636c4979.

After reverting this commit, the internal_metrics (and possibly other
metrics sources) ends up sending its metrics through the metrics batch
buffer twice before the batch is flushed to the sink. This allows the
counter handling in the metrics batch buffer to generate the required
incremental record in the batch, which then persists after the batch is
reset. Without it, the metrics are sent once, no record is added to the
batch, and it is never added later.

Signed-off-by: Bruce Guenter <bruce@timber.io>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>