enhancement(performance): Adjust the strength of `gauge_add` atomics (#6480)
* Adjust the strength of `gauge_add` atomics

This commit adjusts the strength of the `gauge_add` atomics used by
`OpenGauge`. Previous implementation called `emitter` potentially `n^2 / 2`
times, where `n` is the number of participants. Current implementation calls
`emitter` only once after the participant CPU manages to adjust `OpenGauge`
value, though no guarantee is made about the ordering of participant calls to
`emitter`. That is, this commit guarantees that thread `A` will sequence its call
to `emitter` after modification of the internal `gauge` but makes no guarantee
about sequence of thread `B`'s parallel call to its `emitter`, relative to `A`.

If the point of this code is to sequence calls to `emitter` between threads we
can close this out and I'll make a different PR to clarify in the
documentation.

Closes #6477

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Update documentation, introduce spin_loop hint

After clarification in review of the previous commit it was clear that this code
is intended to implement a spinloop with `emitter` in the critical
section. Update this code's documentation to reflect its nature as a spinloop,
possible future avenues for research if this loop plays hell with OS
schedulers.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Correct spelling of 'gague' to 'gauge'

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>