chore(performance): `file-source` testing and benchmarks (#6742)
* Begin rework of `file-source`

This commit is the start of a process to address #6730. The major changes
introduced in this commit so far are application of clippy suggestions and model
checks of the `read_until_with_max_size` test. I have a pretty good idea of how
that function works now and will be introducing benchmarks.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Clippy fixes

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Introduce benchmark

In this commit I have introduce a benchmark for the read_until etc
function. To do this I've had to make it part of the public API of the crate,
but since the crate sits inside a larger project I'm less chuffed about this. I
have fiddled with the test layout some as well.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* end Cargo.toml in newline

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Slim down the feature list slightly, relax bounds

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Reverse 'bytes' upgrade

This crate looks to be challenging to upgrade. Best to do once
tokio is updated in this project.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Use libc on Windows

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* remove unused file

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>