fix(external docs): Fix rendering of buffers/batches sections (#7305)
It was previously rendering ERB which appears directly in the docs
output. Instead, I do the equivalent string interpolation here in cue.

I'm curious if there is a better way to set `_parameter` than I found
which has a bunch of conditions due to the lack of an `else` clause.

I also fixed an issue that the `kafka` sink supports both `max_events`
and `max_bytes` but has a default of `null`. I updated the sink level
cue definition to let these keys be optional rather than assuming `null`
means unsupported.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>