chore: Divvy up vector-core (#7504)
* Divvy up vector-core

This commit extracts 'buffers' into its own crate inside the vector-core
workspace and slims dependencies in both crates. Calling code is unchanged as
the top-level crate re-exports buffers. I intend to move tests and benchmarks of
buffers into the sub-crate in a subsequent PR.

Resolves #7407

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* correct build issues

This commit also removes `lib/derive_is_enum_variant` which is a duplicate of
`matches!` with a slightly different calling convention. Happy to move this into
another PR if folks think it's too big of a swing.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>