chore: Use a forked version of leveldb-sys (#7522)
* Use a forked version of leveldb-sys

The forked version of leveldb-sys patched in here limits leveldb to a maximum of
10 mmap'ed LDB files, reducing the total memory burden of vector. Upstream
leveldb will map 1000 files, so there's quite a difference in memory
usage. Unfortunately this 10 is not configurable.

At the moment the fork is under my personal Github account. I think we ought to
fork to timberio -- we have an archived version -- and pin the patch to a SHA. I
don't have enough juice in the Github org to achieve this.

Resolves #7514

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* switch to timberio/leveldb-sys

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Revert cargo updates

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Remove accidental Cargo.lock

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

Co-authored-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>