Add an explicit +Unpin to our buffers (#7594)
This commit adds an explicit +Unpin to our buffer type definition, introducing
`EventStream` type alias to tidy up some repitition. This is done to support
pull request #7576 where this changed is needed but is a little outside the
scope of that PR. The major change here is in `src/utilization` where I had to
re-implement the select loop to use an explicit stream definition. The `stream!`
macro used previously couldn't make the `+ Unpin` guarantee. It wasn't clear to
me that it could be fixed upstream.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>