chore: Move 'Transform' into core (#7701)
* Move 'Transform' into core

This commit moves the 'Transform' traits into vector-core. This is an extract of
a larger piece of work that moves the whole of `topology` module into core but
which spiraled out of control. Best to break up for review.

The bulk of the change here has to do with `transform_one` as it can no longer
be exposed from core test-only. It's now a free function in the top-level
workspace.

REF #7150

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* `transforms` -> `transform`

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* be strict about transform lengths

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>