chore(tests): Ensure that tracing output is captured (#7764)
Currently tracing output always appears in tests because it side-steps
the print macros that `libtest` hijacks to capture output. Calling
`with_test_writer` swaps to using `print!` and so is captured
appropriately.

I attempted to refactor this a bit but struggled again to make this more
composable due to the fact that Subscribers and Layers are generic over
much of their configuration. I welcome thoughts here though.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>