feat(cli): Add support for --config-dir (#7767)
* feat(cli): Add support for --config-dir

This adds support for passing a configuration directory to look for
configuartion files to supplement the existing support for specifying
individual files. Vector will ignore any files in this directory that
are not one of its known formats (yaml, json, toml).

The motivation for this change is primarily to support a future change
to the way we manage helm charts to, instead of trying to build one
configuration file, instead write a separate file for managed sources
(`managed.yaml`) rather than translating it everything to TOML and
writing `vector.toml`; however it should also more easily enable users
to use a set of configuration files. We do currently support globbing,
but that has the disadvantage of causing a failure on start-up if the
glob doesn't match any files (which I think is correct behavior).

Inspiration for this feature was taken from
[`consul`](https://www.consul.io/docs/agent/options#commandline_options).

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>