fix(sample transform)!: Accept remap conditions (#7864)
* fix(sample transform): Accept remap conditions

The docs for the `sample` transform were updated to use an example remap
condition but the component itself only accepted `check_fields`-style
conditions. This updates the component to allow either condition style
(similar to the `filter` transform).

I updated the tests to just use the new remap style rather than check
both. I think it's fine to rely on the tests in `src/condition` to cover
the differences between conditions.

Fixes: #7863

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>