chore(releasing): Move where the `DEBUG` info is sensed during builds (#7924)
Moves where we include how debug information is injected into the binary
to `build.rs` rather than CI. This feels a bit safer in that it will
check even outside of CI and augment the version string accordingly.

A release build of vector looks like:

```
vector 0.15.0 (x86_64-apple-darwin)
```

That is, no debug info. A debug build of vector looks like:

```
vector 0.15.0 (x86_64-apple-darwin debug=full)
```

We currently just have a binary `debug` or nothing, which we could
switch back to here. I'm not really sure how often `debug` is set to `1`
to only include line info and if that is substantially different in
meaning than including full debug info.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>