chore: Parse Datadog Log agent payloads faster (#8297)
* Parse Datadog Log agent payloads faster

This commit adjusts the way the datadog agent log payloads are parsed, avoiding
the generic `decode_body` and relying on our knowledge of the payload's
structure to avoid the branch misses etc etc inherent in that function. We gain
20Mb/s of parse speed per investigation done in #8263. This commit passes
tests (locally) but I don't have documentation on hand to assert that this
payload is, in fact, what the datadog agent is meant to send.

Resolves #8286

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Expand LogMsg to include full range of datadog agent fields

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Extract `decode_body`, adjust tests

In this commit I've made each field obligatory, removing the `Option`, which
jives with our current understanding of the agent's behavior. This commit
introduces a separate decoding function with quickcheck test to verify, which
makes testing just the decoding step much more direct that previously. Some of
the insertion calls are adjusted to avoid inefficiencies found in `try_insert`,
partially addressed in #8331

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>