chore: Ignore flakey file-source tests (#8424)
* Ignore flakey file-source tests

This commit marks two tests in the file-source as ignore. Both deal with
restarts. We've fiddled with this mechanism quite a bit lately. My argument is
that file-source needs a fundamental rethink -- I did the basic work in
[cernan](https://github.com/postmates/cernan) before `Future` existed -- and
tinkering with restart isn't working.

The test appears to be flakey only on Mac and Windows, our slowest CI machines,
which drives out CI turn-around time. Consider that #8421 has failed CI in two
runs because of one or other of these tests.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Run the previously ignored tests only on Linux

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>