chore(datadog_logs sink)!: Strike text encoding from Datadog logs sink (#8445)
* Strike text encoding from Datadog logs sink

There are two acceptable encodings for datadog logs ingest: JSON and 'text'. The
text encoding is the more limited of the two and does not make sense for end
users to have as a configuration option, considering that JSON ingest is the
more accepted format for ingest.

This commit strikes the 'encoding' option from the sink entirely. This does
simplify the internal logic of the sink -- good -- but removes other features
inherent in the encoding configuration like field exclusion which may or may not
be a problem.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Don't totally strike encoding

This commit re-introduces the encoding notion but only gives a single default
option, Json. Addresses Jesse's feedback.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Docs update

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>