chore: Rate limit defaults reconsidered (#8472)
* Adjust how we set Tower defaults

This is a follow up from PR #8471. In this commit I've removed any default or
near-default TowerRequestConfig instances from `lazy_static` contexts and
shuffled them into their call-site. I have not changed any of the default
overrides.

The largest change here is the removal of `lazy_static` sites. This macro, while
convenient, introduces serialization into vector -- atomic read which is acq/rel
on x86 -- and I'd like it to be rare in the codebase.

For sinks that set an explicit rate limit I have mostly left these intact,
unless I knew better. That is, some sinks set a rate limit because that is a
documented limit and others set it to be higher than the previous default. If
there was no documentation suggesting that the rate limit was arbitrary I
assumed it was intentionally set and left it be.

Resolves #8470

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Remove duplicate config items for generic http sink

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* remove unused import

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>