chore: Small tweaks around the Datadog sink  (#8488)
This collection of commits tweaks a number of small things in the Datadog sink. I have added documentation describing the goals of the tests, avoid going through path iteration code if the log schema fields are identically named and expand the writing space for the encoder.

I had imagined building off this work to introduce more direct serialization -- see https://github.com/timberio/vector/tree/blt-estimate_json_size for a hint of what I was thinking -- but with this commit vector on my machine clocks in at 230Mb/s in the setup described here. I'm satisfied that it's time to move on.

Resolves #8263

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>