enhancement(sinks): Add support for acknowledgements to more HTTP and batch sinks (#8426)
* Add new method `take_finalizers` to `EventMetadata`

The Datadog sinks require the API key to remain in the metadata. This
does not fit with the model of taking all of the metadata. However, the
finalization scheme only requires the list of finalizers to be taken, so
we can scope it down to just that part of the metadata.

* Move metadata handling out of `encode_event` for HttpSink sinks

* Move metadata handling out of `encode_event` for tcp/udp/unix sinks

* Simplify the ServiceLogic framework slightly

* Drop the `Option` around the encoded event finalizers

* Fix clippy issue in statsd sink

Signed-off-by: Bruce Guenter <bruce.guenter@datadoghq.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>