enhancement(sinks): Use variance calculation to better adapt to RTT times in ARC (#8325)
* Add EWMA type with variance calculations

* Use new EwmaVar type in ARC controller (no behavior changes)

* Use the new variance in the ARC calculations

* Mark ARC config as `deny_unknown_fields`

* Rework the ARC test harness to avoid the polling loop

* Adjust ARC default values

* Calculate `reached_limit` *after* starting a request

Previously, the ARC controller determined if the concurrency limit had
been reached before starting a request. In this way, the controller
would only receive an indication that the limit was reached if the sink
tried to send at least one more than the limit. This negatively affects
performance, as the controller will only increase the limit if it was
reached.

This change calculates the value after starting a request, which is when
the indication of how many requests are in flight is changed. This also
drops a mutex lock from the `acquire` code path.

* Make running mean stats math more accurate

Signed-off-by: Bruce Guenter <bruce.guenter@datadoghq.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>