chore(aws_s3 sink): refactor the S3 sink (#8884)
This commit refactors the s3 sink to a more stream oriented model, removing the explicit `Sink` implementation, similar to the work going on in #8825. This commit and that PR are intended to be the "new-style" for sinks. We believe this commit will also address lock-up issues seen in environments with high cardinality partitions into the s3 sink across a disk buffer. 

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>
Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>