chore: Removes `sink::Response` from the Datadog Logs sink (#9172)
* Removes `sink::Response` from the Datadog Logs sink

This commit removes the referenced trait from the sink and from the
`TowerRequestLayer` implementation. Other callers can choose to leave this
constraint in place but it is not required. This change has also allowed me to
clean up the odd Ok/Err split noted in #8825.

I intend to do a similar bit of work to the s3 sinks.

REF #9140

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Use `AsRef` instead of `Into`

The use of `Into` for this purpose had two problems. First and less important,
being able to express `EventStatus` -> `*ApiResponse` is weird. Second and more
important, taking ownership of the API response _happens_ to work for this sink
but is not a guarantee for other sinks.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* dinged by check-events

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>