chore: Allow for localhost soak testing (#9699)
chore: Allow for localhost soak testing

This pull-request introduces localhost soak testing to the vector
project, part of the work to satisfy #9515. The soak introduced here
is log aggregation, that is datadog_agent -> remap ->
datadog_log. Running the soak is discussed in soak/README.md. Soaks
are defined solely in terraform, which I expect will move around some
as #9618 is worked. Vector containers are built locally for now, but
once #9543 is available we can lean on the CI infrastructure to build
containers some.

Closes #9616
Closes #9617

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>