chore: Introduce a 'syslog -> loki' soak (#9729)
Introduce a 'syslog -> loki' soak

This commit does two things. First, it introduces a 'syslog -> loki' soak that
should be helpful in getting #9506 over the line. Second, it introduces
documentation for the process of adding a new soak test and tidies up some of
the terraform.

The new soak can be run like so:

```
> ./soaks/soak.sh syslog_loki deb81750ba3f23d51f421e10178197ae4f74945a 12f93fa87c8546d586e790e7ee519989370ef7f0
```

This runs a comparision of the soak between
deb81750ba3f23d51f421e10178197ae4f74945a (current master as of this commit) and
12f93fa87c8546d586e790e7ee519989370ef7f0 (current tip of #9506 as of this
commit).

Closes #9750

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>