enhancement(vrl): Add `tally_value` function (#9890)
We had originally added `tally` upon a user request even though it is
likely to be supplanted by the upcoming VRL iteration work. However,
they are sensitive to increases in RSS and found that `tally` increased
RSS more than they'd like. Their purposes are simpler in that they just
want to count the number of occurrences of a specific value which can be
done in a more memory efficient way.

Thus here I add `tally_value` to accomplish this. This function is also
a prime candidate to be deleted when VRL iteration is added so I left it
undocumented for now. Once iteration is added we can decide if either of
these functions is worth keeping around and publish them then.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>