enhancement(observability): improve transform utilization metrics (#9828)
This is a relatively naive but straightforward adaptation of our
existing utilization `Stream` wrapper to non-async transforms. The
benefit is that we can start the "waiting" span sooner than the `Stream`
case, as soon as the transform is done processing and begins sending to
its output. This will provide a much more accurate view of how much work
a transform is doing vs being blocked on downstream components.

As part of this, it was beneficial to make the wrappers for function and
fallible function transforms more similar. This includes moving from
combinators to an `async` block for the function variant, and adding the
`ready_chunks` batching to the fallible variant. It'd be great to share
more code here instead of just duplicating, but that felt even more
outside of the scope of this change.

One kinda messy bit of this ended up being the `last_report`
bookkeeping, which is somewhat annoying to have in this path and also
not nearly as precise as the `IntervalStream`-based approach we use in
the `Stream` wrapper`. Ultimately I left it for now, with the assumption
that it should be able to go away as we move towards centralizing the
metric computations. In that case, I'd envision a split between the
timer itself and the reporting mechanism, where we can have a separate
task responsible for polling the state of the timers on a regular
interval.

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>