chore(generator source): Rename generator source to demo_logs source (#9979)
* chore(generator source): Rename generator source to demo source

Following the discussion in #8249 we believe that `demo` is a more
accurate name for this source and that it should not be used for
benchmarking due to the noise added by having Vector generate its own
load. We also feel that making the source suitable for benchmarking
would make it less suitable for demos.

I aliased the source to preserve compatibility. I'll note it in the
release notes when I prepare them for v0.18.0, but not as a breaking
change.

Closes: #8249

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>