chore(vector sink): rewrite vector sink (v2) in the new style (#10148)
closes #9445

The V1 version is already in the new style, so this only modified the v2 version.

As part of this work, the batcher was made even more generic. You can now choose the output type (not necessarily just a Vec), and you have full control of how the output is built. Instead of just appending to a Vec, you can build it incrementally in a lambda, "reducer" style. The batch "data", and "limiter" were also split into respective traits so it's easy to override one or the other. You can see an example of the "reducer" style in the sink.rs of the vector sink.

Signed-off-by: Nathan Fox <fuchsnj@gmail.com>
Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>