chore: Introduce a new http -> pipelines -> blackhole soak (#10142)
* Introduce a new http -> pipelines -> blackhole soak

This commit introduces a new soak to investigate our pipelines in an ongoing
basis. The pipeline configuration for vector and the sample data was contributed
by work was contributed @vladimir-dd. I have expanded the soak to allow mounting
a TESTNAME/data to smuggle static data into the minikube, by which we feed
http_gen. Unfortunately this does not work as we cannot use the Virtual Box VM
for soak testing and https://github.com/kubernetes/minikube/issues/12301 is
open. We will have to figure out some other way to smuggle data into the
minikube for this soak to function.

This commit depends on https://github.com/vectordotdev/vector/pull/10141 which
depends on https://github.com/blt/lading/pull/119.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Use configmap hack

It turns out that the problem discussed in the last commit is fixed by
https://github.com/kubernetes/minikube/pull/13013. This is not present in a
minikube release yet, so we can't rely on it. That said, this commit introduces
a hack whereby the bootstrap is passed in a configmap. This limits the size of
the bootstrap with all the problems that come along with a low-entropy
experiment.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>