chore: Improve character delimited code performance somewhat (#10581)
* Improve character delimited code performance somewhat

This commit improves the performance of the character delimited, byte
deserializing codec somewhat. The major changes to observe here are the use of
memchr for searching for our needle in the haystack instead of a custom loop and
using this to discard over-long sub-buffers, tightening up the loop in that
struct. Included is the program codec_perf that replicates the benchmark loop,
allowing for perf collection. As of this commit the major time sinks are:

* `<vector::codecs::format::bytes::BytesDeserializer as Deserializer>::parse`
* `vector::codecs::framing::character_delimited::CharacterDelimitedDecoder::decode`
* `BytesMut::split_to`
* `alloc::collections::btree::map::entry::VacantEntry<K,V>::insert`

This `decode` is the caller of `split_to`, `parse` calls `LogEvent::insert_flat`
which is, itself, a btree insertion. Timings captured on my system:

* `character_delimited_bytes/no_max`: 114MiB/s
* `character_delimited_bytes/no_max`: 209MiB/s
* `newline_bytes/no_max`: 369MiB/s
* `newline_bytes/no_max`: 1.5GiB/s

REF #10543

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* remove codec_perf test ding

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* test ding

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* feedback

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* warn! more consistently

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>