chore: Address sources of non-vector variability (#10730)
Address sources of non-vector variability

This commit reduces sources of non-vector variability in our soak tests. Lading
is updated here with two key changes: the tcp_gen no longer loses its block
cache when it loses its TCP connection, reducing zero throughput intervals from
that tool and we seed randomness in lading explictly. We make three replica
tests of each experiment rather than two, doing a zero-filter on experiments to
remove samples taken prior to proper startup of the experimental
apparatus. Dependency issues between the soak pods are improved and the control
scripts coordinate more with the in-kube observer to avoid starting the sample
clock before samples are taken.

In practice we still see variability from the minikube environment itself -- I
have observed coredns crashing -- and I've begun to suspect we might avoid this
entirely by adding a new run mode to lading, see this comment:
https://github.com/vectordotdev/vector/issues/10447#issuecomment-1009198851

We now produce graphs of vector throughput, though only as artifacts and not
displayed anywhere. This has bearing on #10494.

I believe this is the more correct solution to #10400, as opposed to
PR #10627. While that PR does function it does not leave us any better off with
regard to debugging, where the changes made here are both easier to diagnose --
the graphs are quite nice -- and the observer is easier to reprogram.

I have removed any explicit erratic tests from the soak workflow as well. We are
now well set up for #10606, the implementation of which will have bearing on
issue #10777 as both are related less to regression detection and single-SHA
analysis.

Resolves #10400
Closes #10627

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>