Shave time off our build, use less noisy build infra (#10810)
This commit shaves at least half time off our container build, runs
soaks on less variable soak instances from a carved up dedicated
host. Previously we saw the soak-vector container take 18 minutes to
fully build. The use of a dedicated building machine shaves half that
time. We have also improved caching rules in the dockerfile itself so,
if the host is primed, the build might be very much faster than the
improved 10 minutes.

We have unrolled the workflow matrix to allow variants to proceed with
better parallelism, unfortunately requiring duplication of mostly
identical blocks in the yml as we can't depend on
sub-matricies. That's the biggest change here, the unrolling of our
computed workflows. Unfortunately there's no ability to `need` on a
sub-matrix which meant we were bottlenecked on _both_ builds
completing. This structure, while annoying, does also allow us to do
erratic detection per variant, with reference to #10606.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>