chore: Ensure we collect at least 200 `bytes_written` samples (#11043)
This commit ensures that we collect a sufficient number of bytes_written samples
from the underlying experiment. In some cases -- such as when minikube restarts
pods -- we can see collection blink out for a time. Previously we collected for a fixed time,
meaning that we would lose sample opportunities. With this change we expose ourselves
to unbounded time runs if a soak test is buggy. The observer now logs somewhat
differently.

I have also bumped the total samples collected back to 200 from 60. This was an accidental
change.

Signed-off-by: Brian L. Troutwine brian@troutwine.us
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>