chore: Remove explicit lazy_static in project (#11243)
* Remove explicit lazy_static in project

This commit removes the use of lazy_static! in favor of OnceCell's Lazy. They
are functionally similar although OnceCell's implementation has better
characteristics in contended situations. This change was motivated by
examination of the profiling results from #11206 where a regex to do with date
parsing was seen before a CPU quiet period. It's unknown if this change will
materially impact that, however we should no longer use lazy_static and so the
time seemed ripe.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* unused ding in vrl/stdlib

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>