chore(performance): emit splunk_hec metrics per request (#11289)
Previously this was done in a streaming fashion, where a single internal
event was emitted with each individual event built. Now we build and
collect all the events, and emit one internal event describing the
total. This also lets us remove the awkward `send_result_stream` method
from `SourceSender`, of which this was the only caller.

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>