chore: Remove trivial instances of getset use (#11598)
* Remove trivial instances of getset use

The crate getset is used in only a few places in the codebase and, in most of
those, does quiet trivial things. This commit removes those spots with the
intention of teeing up a PR to remove getset entirely. We're already seeing that
builds fail on some systems because our dependency (and thus link) list is too
long, hence the desire to remove trivial dependencies where possible.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* more getset removal

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* remove final getset instances

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>