Set mean drift threshold to 7% from 5% in soaks (#11993)
In the past we set the threshold for mean drift at the 9th quartile. Feeling
that this was too high a bar we moved the threshold down to 5%. This, for the
most part, has worked out well, except for some soaks that are, for reasons
unknown, inclined to walk back and forth across that line. 6% is also regularly
violated but 7% rarely is.

In the longer term we'd like to make tolerances here tighter, of course, but
have a need to not force devs to re-run soaks just to see if the results are
accurate or random chance.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>