docs: Audit tls settings (#12046)
* docs: Audit tls settings

I did a sweep of the docs to make sure the `tls` settings lined up with
reality.

As part of this I:

* dropped the `tls.can_enable` cue field since any source that had
  `tls.enabled = true` supports enabling it
* Adds support for the `tls` settings to the `datadog_events` sink.
  These settings were previously being ignored

Note there is a difference here with which components literally support
connecting over TLS and which allow configuration of the `tls` settings.
`tls.enabled` currently indicates which components support configuration
of the settings. I could see us adding another field in the future to
indicate which components support TLS connections. An example of this is
the `redis` source which supports connecting over TLS, but doesn't
support configuration of the `tls` settings.

Closes: #6132

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>