enhancement(buffers): Handle negative acknowledgements by stopping readers (#13563)
This changes the behavior of all sinks that can emit negative acknowledgements. If a sink that is
configured with a disk buffer receives an error or rejection, it will cause the buffer to stop
delivering new events to the sink and emit an error message. This can be resolved through the use of
an additional tool, integrated into vector, that is being developed separately.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>