fix(host_metrics source): Fix handling parsing Linux hybrid cgroups (#13719)
With hybrid cgroups, the `cpu` and `memory` controllers are split
between the base directory (containing the legacy controllers) and the
`unified` subdirectory (containing the modern groups). The `cpu`
controller is usually found in the modern groups, but the top-level
stats are found under the legacy controller in some setups. Similarly,
the `memory` controller can be found in either location.

As such, while autodetecting what mode cgroups are running in (legacy,
hybrid, or modern) is straightforward, detecting if certain controllers
are in use is more complicated. So, this rewrites the cgroup metrics
parsing to recurse into all of the relevant directories, scan for the
cpu and memory data files, and emit metrics for any that are found. This
unfortunately prevents us from emitting the warning about the memory
controller not being enabled, but does give coverage for actually
emitting the metrics for any memory controller configuration.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>