chore(observability): Remove extraneous to_string (#13749)
* chore(observability): Remove extraneous to_string and fix mode naming

The socket sink/source related metrics had an extraneous `.to_string()` on some
of the constant strings used for modes, which causes an unnecessary memory
allocation.

While making the change, I noticed that some of the events emit this value as
`mode` and some as `protocol`, so I changed them all to `mode` which is more
common for this value.

* Restore the original tag naming

This results in the inconsistency where the `BytesReceived` and `BytesSent`
events emit traces and metrics with the `protocol` tag, and the `EventsReceived`
and `EventsSent` events emit with the `mode` tag. The former is required by the
component spec, but the latter is the current convention for these socket
events.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>