revert(observability): Convert internal metric counts back to absolute (#14251)
Revert "enhancement(observability): Convert internal metric counters to incremental (#13883)"

This reverts commit 15540a3167daa18d46de067265e72cec581261cd.

The above change broke metrics exposed in `vector top` and the API, as they
assumed the values were absolute. Also, the current setup does not work in the
presence of multiple consumers pulling from the metrics registry, as the
increments will be split across the consumers with none seeing the whole value.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>