chore(dev): Restore the `-rdynamic` linker flag (#14326)
In order for dynamically loaded Lua modules to work, they need to be able to
link against the symbols provided by the Lua libraries. These symbols are not
exposed unless `vector` is linked with the `-rdynamic` flag. This flag was
removed by #13074 (Fix application of `rustflags`) and here we are restoring
just this flag for linking.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>