fix(observability): Skip expiring registered metrics (#14489)
Registered internal metrics interact badly with the newly released expiry feature. If
an internal metric is registered but not updated for longer than the expiry interval,
the `Arc` handle becomes disconnected from the registry. Subsequent updates to
that handle no longer have any effect on the registry, and the value never re-appears.

* Import the `recency` module from `metrics_util`

* Modify recency tracking to account for outstanding handles

We don't want to expire internal metrics if there is an outstanding handle that
could later update the shared value. So, here we look up the count of references
to the inner value to see if there are more than expected.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>