fix(tag_cardinality_limit transform): Handle drop event edge case (#14889)
* fix(tag_cardinality_limit transform): Handle drop event edge case

With a limit of 2 and the action set to drop_event, given the following two
sequences of tags:

1. tag1=>val1,tag2=>val1
2. tag1=>val2,tag2=>val1
3. tag1=>val3,tag2=>val2
4. tag1=>val1,tag2=>val3

1. tag1=>val1,tag2=>val1
2. tag1=>val1,tag2=>val2
3. tag1=>val2,tag2=>val3
4. tag1=>val3,tag2=>val1

We should have identical behavior in both cases: event 3 should be rejected
because a tag went over the limit, but event 4 should be accepted because the
tags emitted by this transform are still within the limits.

The above behavior is not the current behavior. In the first scenario, only
events 3 would be rejected because tag2 would not be recorded, and in the latter
scenario, both events 3 and 4 would be rejected. This PR corrects the
inaccuracy.

* Use `hashbrown` to avoid double and triple lookups on tag names

* Drop redundant use of Cow

* Simplify TagValueSet::new
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>