feat(ci): add analysis from smp to regression workflow (#15063)
* fix(ci): add --team-id arg to `smp job cancel`

All of the `smp` subcommands take a `--team-id` argument between `smp`
and `job`, but this flag is missing in the `smp job cancel` invocation
in the Vector Regression workflow, so this commit adds that flag.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* chore(ci): update submission flags for smp 0.5.0

Flags for the `smp job submit` step have changed: as a result of
passing in image tags to specify the baseline and comparison variants
of a target in regression testing, `smp` no longer has access to Git
commit SHA metadata via the `--baseline-image` and
`--comparison-image` flag parameters. To enable `smp` to report this
information during its analysis step, the `--baseline-sha` and
`--comparison-sha` flags were added to `smp job submit`. This commit
adds those flags and the corresponding commit hashes to the job
submission step in the regression workflow.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* feat(ci): add smp analysis to regression workflow

To add regression reports from `smp`, this commit uncomments and
renames the analysis job, making minor tweaks so that the resulting
job:

    * Downloads an `smp` binary with the right version.

    * Uses that binary to download an HTML report analyzing regression
      experiments on a Vector PR.

    * Posts that HTML report into a comment on that PR.

    * Uploads the HTML report to workflow run artifacts.

Hat-tip and thanks to Brian L Troutwine & George Hahn for making these
changes easy.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* fix(ci): add missing dependency to analysis job

This commit adds a missing task dependency to the analysis job so that
it downloads the correct version of `smp`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* fix(ci): use configure-aws-credentials@v1-node16

GitHub Actions CI emits warnings about the
`aws-actions/configure-aws-credentials@v1` action because it uses
Node.js 12, and Node.js 12 actions are deprecated with removal set for
Summer 2023. Migrating to Node.js-16-based GitHub Actions is the
suggested mitigation, and the
`aws-actions/configure-aws-credentials@v1-node16` is suggested as a
replacement, so this commit makes that change.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* fix(ci): give build-comparison more accurate name

In the Regression task, the `build-comparison` job, which builds a
Vector container for the comparison variant of a regression
experiment, uses the display name "Build baseline Vector
container". This name confused me when I first saw the output in
GitHub Actions because I thought I might have introduced a bug into
the CI Regression workflow, but it appears that the `build-comparison`
job is correctly configured aside from having a misleading name.

To avoid this sort of confusion in the future, this commit updates the
display name of the `build-comparison` job in the Regression workflow
to reflect more accurately what that job does.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* add region for s3

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Enable acks in splunk_hec

It turns out that lading will crash on the splunk_hec_route_s3 experiment
because it expects acks. We have to intentionally inform Vector to ack.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* no-ack variant intended

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>
Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>
Co-authored-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>